Small cleanup

+++ rpcs3/Emu/SysCalls/lv2/sys_fs.cpp
@@ -20,7 +20,7 @@
 
 SysCallBase sys_fs("sys_fs");
 
-s32 sys_fs_open(vm::ptr<const char> path, s32 flags, vm::ptr<u32> fd, u64 mode, vm::ptr<const void> arg, u64 size)
+s32 sys_fs_open(vm::ptr<const char> path, s32 flags, vm::ptr<u32> fd, u32 mode, vm::ptr<const void> arg, u64 size)
 {
 	sys_fs.Warning("sys_fs_open(path=*0x%x, flags=0x%x, fd=*0x%x, arg=*0x%x, size=0x%llx)", path, flags, fd, arg, size);
 	sys_fs.Warning("*** path = '%s'", path.get_ptr());
